backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
bcrypt==4.1.3
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
pydantic>=2.6.4
